+++ Lidgren.Network/NetPeer.Send.cs
@@ -172,7 +172,7 @@ namespace Lidgren.Network
 			msg.m_messageType = NetMessageType.Unconnected;
 
 			Interlocked.Increment(ref msg.m_recyclingCount);
-			m_unsentUnconnectedMessages.Enqueue(new NetTuple<NetEndPoint, NetOutgoingMessage>(new NetEndPoint(adr, port), msg));
+			m_unsentUnconnectedMessages.Enqueue((new NetEndPoint(adr, port), msg));
 		}
 
 		/// <summary>
@@ -195,7 +195,7 @@ namespace Lidgren.Network
 			msg.m_isSent = true;
 
 			Interlocked.Increment(ref msg.m_recyclingCount);
-			m_unsentUnconnectedMessages.Enqueue(new NetTuple<NetEndPoint, NetOutgoingMessage>(recipient, msg));
+			m_unsentUnconnectedMessages.Enqueue((recipient, msg));
 		}
 
 		/// <summary>
@@ -227,7 +227,7 @@ namespace Lidgren.Network
 
 			Interlocked.Add(ref msg.m_recyclingCount, recipients.Count);
 			foreach (NetEndPoint ep in recipients)
-				m_unsentUnconnectedMessages.Enqueue(new NetTuple<NetEndPoint, NetOutgoingMessage>(ep, msg));
+				m_unsentUnconnectedMessages.Enqueue((ep, msg));
 		}
 
 		/// <summary>